Fixed overflow bug with alphabetic_file_to_numeric().
Instead of panicking when we overflow, we delay failure of the overflow
until the *next* attempted calculation using the overflowed power. This
will still fail if the file string is too long/large, but I'm okay with
that, it should be fine for a normal chess board and can still
theoretically be expanded to larger boards.