Fixed regex for parsing moves.
Discriminant and destination regex are now limited to a-h letters. We
originally wanted to allow for all alphabetical characters for larger
boards, but this complicates things, the x for captures ends up being
treated as if it were a file name. This change simplifies the regex and
makes it work better. Some tests that were failing pass now.