Added bitboard struct
Added documentation for bitboards as well as initial methods for
converting from 2D board coordinates to 1D bitboard index. Added error
type for out-of-bounds and added tests for coordinate conversions. Also
made Square rank and file public. Maybe using getters or setters for
this would be better, but for now I think it's fine that the square
coordinates are public.