Did some benchmarking to test bitmask variation performance
A for loop is slightly too complex for Rust's compiler (even optimized)
to simplify and hard-code. A match expression with hard coded bitmasks
for each file is about 10x as performant as other implementations. So
despite being a little more work to write the function, I'll probably
use the match implementation.